Frontend consumes X-Script-Name request header to correct paths
This fixes #3, although it introduces problems of its own when trying to
get the routing working. Specifically, we can't know the base path when
no request is involved (i.e., the client-side JavaScript) and as such,
we have to deconstruct the route handler, looking at the last section,
and matching that with known routes. This will overgeneralise if the
last part of the base path is *also* a known route. For example:

    Base path      /app/contacts

    Home route     /app/contact         => contacts (FIXME)
    About route    /app/contact/about   => about
    Contact route  /app/contact/contact => contact
    etc.

A potentially more robust approach would be to hardcode the base path in
the `.env` file and not mess about with request headers at all...