Commented out containerisation stuff, temporarily
The Docker containerisation development has been put on the back burner.
It works fine, in principle, but needs to be fleshed out once the whole
application stack is more functional. The work already done is staged in
the `docker` branch -- this README was missed -- to be enhanced and
merged into `master` at a later date...